delete oauth. OAuthTokenService renamed to AuthTokenService. Authentication only by password and login
