+++ devicehive-frontend/src/main/java/com/devicehive/application/JerseyConfig.java
@@ -42,7 +42,6 @@ public class JerseyConfig extends ResourceConfig {
 
         registerClasses(ApiInfoResourceImpl.class,
                 ConfigurationResourceImpl.class,
-                AuthJwtTokenResourceImpl.class,
                 DeviceClassResourceImpl.class,
                 DeviceCommandResourceImpl.class,
                 DeviceNotificationResourceImpl.class,
