--- devicehive-frontend/src/main/java/com/devicehive/model/oauth/AuthProvider.java
@@ -1,36 +1,0 @@
-package com.devicehive.model.oauth;
-
-/*
- * #%L
- * DeviceHive Java Server Common business logic
- * %%
- * Copyright (C) 2016 DataArt
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-import com.devicehive.vo.OauthJwtRequestVO;
-import com.devicehive.vo.JwtTokenVO;
-
-import javax.validation.constraints.NotNull;
-
-/**
- * Provides JWT Token from OAuth2 authorization
- */
-public abstract class AuthProvider {
-
-    public abstract boolean isIdentityProviderAllowed();
-
-    public abstract JwtTokenVO createAccessKey(@NotNull final OauthJwtRequestVO accessKeyRequest);
-}
